fix: homogenize shielding component
