+++ apps/namadillo/src/App/Masp/MaspInternalShield.tsx
@@ -1,0 +1,183 @@
+import { Chain } from "@chain-registry/types";
+import { AccountType } from "@namada/types";
+import { params } from "App/routes";
+import {
+  OnSubmitTransferParams,
+  TransferModule,
+} from "App/Transfer/TransferModule";
+import { allDefaultAccountsAtom } from "atoms/accounts";
+import { namadaTransparentAssetsAtom } from "atoms/balance";
+import { chainParametersAtom } from "atoms/chain";
+import { ledgerStatusDataAtom } from "atoms/ledger";
+import { rpcUrlAtom } from "atoms/settings";
+import BigNumber from "bignumber.js";
+import { useTransactionActions } from "hooks/useTransactionActions";
+import { useTransfer } from "hooks/useTransfer";
+import { useUrlState } from "hooks/useUrlState";
+import { wallets } from "integrations";
+import invariant from "invariant";
+import { useAtom, useAtomValue } from "jotai";
+import { createTransferDataFromNamada } from "lib/transactions";
+import { useState } from "react";
+import namadaChain from "registry/namada.json";
+
+export const MaspInternalShield = ({
+  type,
+}: {
+  type: "shield" | "unshield";
+}): JSX.Element => {
+  const { storeTransaction } = useTransactionActions();
+  const [displayAmount, setDisplayAmount] = useState<BigNumber | undefined>();
+  const [generalErrorMessage, setGeneralErrorMessage] = useState("");
+  const [currentStatus, setCurrentStatus] = useState("");
+  const [currentStatusExplanation, setCurrentStatusExplanation] = useState("");
+
+  const rpcUrl = useAtomValue(rpcUrlAtom);
+  const chainParameters = useAtomValue(chainParametersAtom);
+  const defaultAccounts = useAtomValue(allDefaultAccountsAtom);
+  const [ledgerStatus, setLedgerStatusStop] = useAtom(ledgerStatusDataAtom);
+  const { data: availableAssets, isLoading: isLoadingAssets } = useAtomValue(
+    namadaTransparentAssetsAtom
+  );
+
+  const chainId = chainParameters.data?.chainId;
+  const sourceAddress =
+    type === "unshield" ?
+      defaultAccounts.data?.find(
+        (account) => account.type !== AccountType.ShieldedKeys
+      )?.address
+    : defaultAccounts.data?.find(
+        (account) => account.type === AccountType.ShieldedKeys
+      )?.address;
+  const destinationAddress =
+    type === "shield" ?
+      defaultAccounts.data?.find(
+        (account) => account.type === AccountType.ShieldedKeys
+      )?.address
+    : defaultAccounts.data?.find(
+        (account) => account.type !== AccountType.ShieldedKeys
+      )?.address;
+
+  const [selectedAssetAddress, setSelectedAssetAddress] = useUrlState(
+    params.asset
+  );
+  const selectedAsset =
+    selectedAssetAddress ? availableAssets?.[selectedAssetAddress] : undefined;
+
+  const ledgerAccountInfo = ledgerStatus && {
+    deviceConnected: ledgerStatus.connected,
+    errorMessage: ledgerStatus.errorMessage,
+  };
+
+  const {
+    execute: performTransfer,
+    isPending: isPerformingTransfer,
+    isSuccess,
+    error,
+    txKind,
+    feeProps,
+    completedAt,
+    redirectToTransactionPage,
+  } = useTransfer({
+    source: sourceAddress ?? "",
+    target: destinationAddress ?? "",
+    token: selectedAsset?.originalAddress ?? "",
+    displayAmount: displayAmount ?? new BigNumber(0),
+    onUpdateStatus: setCurrentStatus,
+    onBeforeBuildTx: () => {
+      setCurrentStatus("Generating MASP Parameters...");
+      setCurrentStatusExplanation(
+        "Generating MASP parameters can take a few seconds. Please wait..."
+      );
+    },
+    onBeforeSign: () => {
+      setCurrentStatus("Waiting for signature...");
+    },
+    onBeforeBroadcast: async () => {
+      setCurrentStatus("Broadcasting Shielding transaction...");
+    },
+    onError: async () => {
+      setCurrentStatus("");
+      setCurrentStatusExplanation("");
+    },
+    asset: selectedAsset?.asset,
+  });
+
+  // We stop the ledger status check when the transfer is in progress
+  setLedgerStatusStop(isPerformingTransfer);
+
+  const onSubmitTransfer = async ({
+    memo,
+  }: OnSubmitTransferParams): Promise<void> => {
+    try {
+      setGeneralErrorMessage("");
+      setCurrentStatus("");
+
+      invariant(sourceAddress, "Source address is not defined");
+      invariant(chainId, "Chain ID is undefined");
+      invariant(selectedAsset, "No asset is selected");
+
+      const txResponse = await performTransfer({ memo });
+
+      if (txResponse) {
+        const txList = createTransferDataFromNamada(
+          txKind,
+          selectedAsset.asset,
+          rpcUrl,
+          true,
+          txResponse,
+          memo
+        );
+
+        // Currently we don't have the option of batching transfer transactions
+        if (txList.length === 0) {
+          throw "Couldn't create TransferData object";
+        }
+        const tx = txList[0];
+        storeTransaction(tx);
+      } else {
+        throw "Invalid transaction response";
+      }
+    } catch (err) {
+      setGeneralErrorMessage(err + "");
+    }
+  };
+
+  return (
+    <TransferModule
+      source={{
+        isLoadingAssets,
+        availableAssets,
+        selectedAssetAddress,
+        availableAmount: selectedAsset?.amount,
+        chain: namadaChain as Chain,
+        availableWallets: [wallets.namada],
+        wallet: wallets.namada,
+        walletAddress: sourceAddress,
+        isShieldedAddress: type === "unshield",
+        onChangeSelectedAsset: setSelectedAssetAddress,
+        amount: displayAmount,
+        onChangeAmount: setDisplayAmount,
+        ledgerAccountInfo,
+      }}
+      destination={{
+        chain: namadaChain as Chain,
+        availableWallets: [wallets.namada],
+        wallet: wallets.namada,
+        walletAddress: destinationAddress,
+        isShieldedAddress: type === "shield",
+      }}
+      feeProps={feeProps}
+      isSubmitting={isPerformingTransfer || isSuccess}
+      errorMessage={generalErrorMessage || error?.message}
+      currentStatus={currentStatus}
+      currentStatusExplanation={currentStatusExplanation}
+      onSubmitTransfer={onSubmitTransfer}
+      completedAt={completedAt}
+      buttonTextErrors={{
+        NoAmount: "Define an amount to shield",
+      }}
+      onComplete={redirectToTransactionPage}
+    />
+  );
+};
